requirements.txt
# Core dependencies
vk_api==1.3.5
requests==2.31.0
aiohttp==3.9.1

# RSS и парсинг
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3

# База данных
aiosqlite==0.19.0

# Telegram бот
python-telegram-bot==20.7

# ИИ и обработка текста
# Ollama клиент для локальной LLM
ollama==0.1.6

# Утилиты
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Асинхронность
asyncio==3.4.3

# Хеширование и работа с данными
hashlib-compat==1.0.0

# Прокси и ротация
fake-useragent==1.4.0

# Логирование
loguru==0.7.2

# Тестирование (опционально)
pytest==7.4.3
pytest-asyncio==0.21.1
